code-server-backend/requirements.txt
uvicorn
fastapi
pydantic
prisma
python-dotenv
colorlog
docker